+++ CHANGELOG.md
@@ -5,11 +5,26 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][unreleased]
 ### Changed
 
-## [0.0.2] - 2015-06-02
+## [0.3.0] - 2015-06-04
+### Changed
+- controllers/Discussions.php - Added permission checks.
+- controllers/Comments.php - Added permission checks.
+- controllers/Users.php - Added permission checks.
+- language/english/forums_lang.php - Added new error message for permissions system.
+
+### Added
+- libraries/Permission.php - Added a simple group based permissions system.
+- models/Install_m.php - Added more sql to create permissions and permission_map table along with default data.
+
+
+## [0.2.0] - 2015-06-02
 ### Changed
 - core/MY_Controller, changed version number for minor feature additions.
 - language/english/forums_lang.php, added additional language files for new pages.
 - models/Install_m.php, updated the users table creation.
+- models/Install_m.php, updated language packs sql to include italian by default.
+- controllers/Install.php, Updated the install process to allow users to pick the sites default language.
+- templates/install/settings.php, Added the language dropdown box.
 
 ### Added
 - templates/default/pages/users/profile.php